added exposition and a general engine
exposition is handled as individual scenes with the "exposition" value as a header. I wrote a general engine controlled from the ntv file and completely generalised in terms of questions and how many events and what not, also added a "question" tag to ntv files that takes a question type string whihc is passed to the question class where it can be interprited into a which question to ask. Also fixed the engine thread hanging around after killing tkinter window using os._exit()